third_party/WebKit/LayoutTests/http/tests/inspector/bindings/shadowdom-bindings.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>

async function test()
{
    InspectorTest.markStep('dumpInitialWorkspace');
    var snapshot = InspectorTest.dumpWorkspace();

    InspectorTest.markStep('attachShadow1');
    await Promise.all([
        InspectorTest.attachShadowDOM('shadow1', '#template', '_test_attachShadow1.js'),
        InspectorTest.waitForSourceMap('sourcemap-script.js.map'),
        InspectorTest.waitForSourceMap('sourcemap-style.css.map'),
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('attachShadow2');
    await Promise.all([
        InspectorTest.attachShadowDOM('shadow2', '#template', '_test_attachShadow2.js'),
        InspectorTest.waitForSourceMap('sourcemap-script.js.map'),
        InspectorTest.waitForSourceMap('sourcemap-style.css.map'),
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachShadow1');
    await Promise.all([
        InspectorTest.detachShadowDOM('shadow1', '_test_detachShadow1.js'),
        InspectorTest.waitForStyleSheetRemoved('sourcemap-style.css'),
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachShadow2');
    await Promise.all([
        InspectorTest.detachShadowDOM('shadow2', '_test_detachShadow2.js'),
        InspectorTest.waitForStyleSheetRemoved('sourcemap-style.css'),
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.completeTest();

}

</script>

</head>
<body onload="runTest()">
<p>
Verify that UISourceCodes are added and removed as shadow dom styles and scripts are added and removed.
</p>

<template id='template'>
<style>div {
    color: blue;
}
/*# sourceURL=sourcemap-style.css */
/*# sourceMappingURL=resources/sourcemap-style.css.map */
</style>
<script>window.foo = console.log.bind(console, 'foo');
//# sourceURL=sourcemap-script.js
//# sourceMappingURL=resources/sourcemap-script.js.map
</script>
<p>Hi! I'm ShadowDOM v1!</p>
</template>

</body>
</html>
