third_party/WebKit/LayoutTests/http/tests/inspector/bindings/shadowdom-bindings-expected.txt

Verify that UISourceCodes are added and removed as shadow dom styles and scripts are added and removed.


Running: dumpInitialWorkspace
Removed: 0 uiSourceCodes
Workspace: 7 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/inspector-test.js

Running: attachShadow1
Removed: 0 uiSourceCodes
Workspace: 14 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] shadowdom-bindings.html
[+] debugger:///VM[XXX] sourcemap-script.js
[+] debugger:///VM[XXX] _test_attachShadow1.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
[+] http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-sass.scss
[+] http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-typescript.ts
    http://127.0.0.1:8000/inspector/bindings/shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/inspector-test.js
[+] sourcemap-script.js
[+] sourcemap-style.css
[+] _test_attachShadow1.js

Running: attachShadow2
Removed: 0 uiSourceCodes
Workspace: 17 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] sourcemap-script.js
[+] debugger:///VM[XXX] sourcemap-script.js
    debugger:///VM[XXX] _test_attachShadow1.js
[+] debugger:///VM[XXX] _test_attachShadow2.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-sass.scss
    http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-typescript.ts
    http://127.0.0.1:8000/inspector/bindings/shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/inspector-test.js
    sourcemap-script.js
    sourcemap-style.css
    _test_attachShadow1.js
[+] _test_attachShadow2.js

Running: detachShadow1
Removed: 0 uiSourceCodes
Workspace: 19 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] sourcemap-script.js
    debugger:///VM[XXX] sourcemap-script.js
    debugger:///VM[XXX] _test_attachShadow1.js
    debugger:///VM[XXX] _test_attachShadow2.js
[+] debugger:///VM[XXX] _test_detachShadow1.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-sass.scss
    http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-typescript.ts
    http://127.0.0.1:8000/inspector/bindings/shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/inspector-test.js
    sourcemap-script.js
    sourcemap-style.css
    _test_attachShadow1.js
    _test_attachShadow2.js
[+] _test_detachShadow1.js

Running: detachShadow2
Removed: 2 uiSourceCodes
[-] http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-sass.scss
[-] sourcemap-style.css
Workspace: 19 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] shadowdom-bindings.html
    debugger:///VM[XXX] sourcemap-script.js
    debugger:///VM[XXX] sourcemap-script.js
    debugger:///VM[XXX] _test_attachShadow1.js
    debugger:///VM[XXX] _test_attachShadow2.js
    debugger:///VM[XXX] _test_detachShadow1.js
[+] debugger:///VM[XXX] _test_detachShadow2.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-typescript.ts
    http://127.0.0.1:8000/inspector/bindings/shadowdom-bindings.html
    http://127.0.0.1:8000/inspector/inspector-test.js
    sourcemap-script.js
    _test_attachShadow1.js
    _test_attachShadow2.js
    _test_detachShadow1.js
[+] _test_detachShadow2.js
